test(telegram): extract isolated bot fixtures
